Move Bytes and BytesVectored trait to new bytes module
As they are not limited to networking.
